addresses PR comments; renames files properly;
